+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -1,8 +1,7 @@
-package domain;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.board.ChessBoard;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
