+++ src/test/java/chess/domain/piece/QueenTest.java
@@ -1,9 +1,7 @@
-package domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.piece.Color;
-import chess.domain.piece.Queen;
 import chess.domain.position.File;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
