+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,10 +1,7 @@
-package domain;
+package chess.domain.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.position.File;
-import chess.domain.position.Position;
-import chess.domain.position.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
