+++ src/test/java/chess/domain/position/PositionsTest.java
@@ -1,11 +1,7 @@
-package domain;
+package chess.domain.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.position.File;
-import chess.domain.position.Position;
-import chess.domain.position.Positions;
-import chess.domain.position.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
